# lib/api-zod/src/generated/types/apiErrorResponseDto.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Med-Ease enterprise healthcare platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ApiErrorResponseDto {
  status: string;
  statusCode: number;
  /** Error message or validation errors */
  message: string;
  path: string;
  timestamp: Date;
}